refactor(services): move the logic to the services
