+++ src/Infrastructure/Repositories/Interfaces/ICartRepository.cs
@@ -7,57 +7,66 @@ namespace Tienda_UCN_api.Src.Infrastructure.Repositories.Interfaces
     /// </summary>
     public interface ICartRepository
     {
-        /// <summary>
-        /// Crea un nuevo carrito o devuelve uno existente.
-        /// </summary>
-        /// <param name="buyerId">ID del comprador.</param>
-        /// <param name="userId">ID del usuario si es que está autenticado</param>
-        /// <returns>Tarea que representa la operación asincrónica retornando un objeto Cart.</returns>
-        Task<Cart> CreateOrGetAsync(string buyerId, int? userId = null);
 
         /// <summary>
-        /// Limpia el carrito de compras.
+        /// Busca un carrito de compras.
         /// </summary>
         /// <param name="buyerId">ID del comprador.</param>
         /// <param name="userId">ID del usuario si es que está autenticado</param>
         /// <returns>Tarea que representa la operación asincrónica retornando un objeto Cart.</returns>
-        Task<Cart> ClearAsync(string buyerId, int? userId = null);
+        Task<Cart?> FindAsync(string buyerId, int? userId);
+
+        /// <summary>
+        /// Obtiene un carrito por ID de usuario.
+        /// </summary>
+        /// <param name="userId">ID del usuario.</param>
+        /// <returns>Tarea que representa la operación asincrónica retornando un objeto Cart.</returns>
+        Task<Cart?> GetByUserIdAsync(int userId);
+
+        /// <summary>
+        /// Obtiene un carrito anónimo.
+        /// </summary>
+        /// <param name="buyerId">ID del comprador.</param>
+        /// <returns>Tarea que representa la operación asincrónica retornando un objeto Cart.</returns>
+        Task<Cart?> GetAnonymousAsync(string buyerId);
+
+        /// <summary>
+        /// Crea un nuevo carrito de compras.
+        /// </summary>
+        /// <param name="buyerId">ID del comprador.</param>
+        /// <param name="userId">ID del usuario si es que está autenticado</param>
+        /// <returns>Tarea que representa la operación asincrónica retornando un objeto Cart.</returns>
+        Task<Cart> CreateAsync(string buyerId, int? userId = null);
+
+        /// <summary>
+        /// Actualiza el carrito de compras.
+        /// </summary>
+        /// <param name="cart">El carrito a actualizar.</param>
+        /// <returns>Tarea que representa la operación asincrónica.</returns>
+        Task UpdateAsync(Cart cart);
+
+        /// <summary>
+        /// Elimina el carrito de compras.
+        /// </summary>
+        /// <param name="cart">El carrito a eliminar.</param>
+        /// <returns>Tarea que representa la operación asincrónica.</returns>
+        /// <param name="cart">El carrito a eliminar.</param>
+        /// <returns>Tarea que representa la operación asincrónica.</returns>
+        Task DeleteAsync(Cart cart);
 
         /// <summary>
         /// Agrega un artículo al carrito de compras.
         /// </summary>
-        /// <param name="buyerId">ID del comprador.</param>
-        /// <param name="productId">ID del producto.</param>
-        /// <param name="quantity">Cantidad del producto.</param>
-        /// <param name="userId">ID del usuario si es que está autenticado</param>
-        /// <returns>Tarea que representa la operación asincrónica retornando un objeto Cart.</returns>
-        Task<Cart> AddItemAsync(string buyerId, int productId, int quantity, int? userId = null);
+        /// <param name="cart">El carrito al que se agregará el artículo.</param>
+        /// <param name="cartItem">El artículo del carrito a agregar.</param>
+        /// <returns>Tarea que representa la operación asincrónica.</returns>
+        Task AddItemAsync(Cart cart, CartItem cartItem);
 
         /// <summary>
         /// Elimina un artículo del carrito de compras.
         /// </summary>
-        /// <param name="buyerId">ID del comprador.</param>
-        /// <param name="productId">ID del producto.</param>
-        /// <param name="userId">ID del usuario si es que está autenticado</param>
-        /// <returns>Tarea que representa la operación asincrónica retornando un objeto Cart.</returns>
-        Task<Cart> RemoveItemAsync(string buyerId, int productId, int? userId = null);
-
-        /// <summary>
-        /// Asocia un carrito de compras con un usuario.
-        /// </summary>
-        /// <param name="buyerId">ID del comprador.</param>
-        /// <param name="userId">ID del usuario.</param>
+        /// <param name="cartItem">El artículo del carrito a eliminar.</param>
         /// <returns>Tarea que representa la operación asincrónica.</returns>
-        Task AssociateWithUserAsync(string buyerId, int userId);
-
-        /// <summary>
-        /// Actualiza la cantidad de un artículo en el carrito de compras.
-        /// </summary>
-        /// <param name="buyerId">ID del comprador.</param>
-        /// <param name="productId">ID del producto.</param>
-        /// <param name="quantity">Nueva cantidad del producto.</param>
-        /// <param name="userId">ID del usuario si es que está autenticado</param>
-        /// <returns>Tarea que representa la operación asincrónica retornando un objeto Cart.</returns>
-        Task<Cart> UpdateItemQuantityAsync(string buyerId, int productId, int quantity, int? userId = null);
+        Task RemoveItemAsync(CartItem cartItem);
     }
 }
